Move StaticPtr into a shared mod
